requirements.txt
django==4.1.6
django-phonenumber-field[phonenumbers]==7.0.2

python-dotenv==0.21.1
